+++ lib/tensorflow/keras/datasets/boston_housing.rb
@@ -1,0 +1,30 @@
+module TensorFlow
+  module Keras
+    module Datasets
+      module BostonHousing
+        def self.load_data(path: "boston_housing.npz", test_split: 0.2, seed: 113)
+          # TODO verify hash
+          data = Utils.load_dataset(
+            path,
+            "https://storage.googleapis.com/tensorflow/tf-keras-datasets/boston_housing.npz"
+          )
+
+          x = data["x"]
+          y = data["y"]
+
+          len = x.shape[0]
+          indices = (0...len).to_a.shuffle(random: Random.new(seed))
+          x = x[indices, true]
+          y = y[indices]
+
+          x_train = x[0...(len * (1 - test_split)), true]
+          y_train = y[0...(len * (1 - test_split))]
+          x_test = x[(len * (1 - test_split))..-1, true]
+          y_test = y[(len * (1 - test_split))..-1]
+
+          [[x_train, y_train], [x_test, y_test]]
+        end
+      end
+    end
+  end
+end